fixing template to work with older ios sdks

+++ setup/TemplateConfigurator.rb
@@ -1,3 +1,4 @@
+require 'fileutils'
 require 'colored'
 
 module Pod
@@ -69,13 +70,22 @@ module Pod
     def run
       @message_bank.welcome_message
 
-      ConfigureIOS.perform(configurator: self)
+      framework = self.ask_with_answers("What language do you want to use?", ["ObjC", "Swift"]).to_sym
+      case framework
+        when :swift
+          ConfigureSwift.perform(configurator: self)
+        
+        when :objc
+          ConfigureIOS.perform(configurator: self)
+      end
+      
 
       replace_variables_in_files
       clean_template_files
       rename_template_files
       add_pods_to_podfile
       customise_prefix
+      ensure_carthage_compatibility
       reinitialize_git_repo
       run_pod_install
 
@@ -84,6 +94,10 @@ module Pod
 
     #----------------------------------------#
 
+    def ensure_carthage_compatibility
+      FileUtils.ln_s('Example/Pods/Pods.xcodeproj', '_Pods.xcodeproj')
+    end
+
     def run_pod_install
       puts "\nRunning " + "pod install".magenta + " on your new library."
       puts ""
@@ -91,6 +105,9 @@ module Pod
       Dir.chdir("Example") do
         system "pod install"
       end
+
+      `git add Example/#{pod_name}.xcodeproj/project.pbxproj`
+      `git commit -m "Initial commit"`
     end
 
     def clean_template_files
@@ -113,6 +130,13 @@ module Pod
       end
     end
 
+    def set_platform_ios7
+      template_podfile_path = "templates/ios/Example/Podfile"
+      text = File.read(template_podfile_path)
+      text.gsub!('use_frameworks!', "platform :ios,'7.0'")
+      File.open(template_podfile_path, "w") { |file| file.puts text }
+    end
+
     def add_pod_to_podfile podname
       @pods_for_podfile << podname
     end
@@ -132,30 +156,32 @@ module Pod
 
     def customise_prefix
       prefix_path = "Example/Tests/Tests-Prefix.pch"
+      return unless File.exists? prefix_path
+      
       pch = File.read prefix_path
       pch.gsub!("${INCLUDED_PREFIXES}", @prefixes.join("\n  ") )
       File.open(prefix_path, "w") { |file| file.puts pch }
     end
 
-    def set_test_framework(test_type)
-      content_path = "setup/test_examples/" + test_type + ".m"
-      tests_path = "templates/ios/Example/Tests/Tests.m"
+    def set_test_framework(test_type, extension)
+      content_path = "setup/test_examples/" + test_type + "." + extension
+      folder = extension == "m" ? "ios" : "swift"
+      tests_path = "templates/" + folder + "/Example/Tests/Tests." + extension
       tests = File.read tests_path
       tests.gsub!("${TEST_EXAMPLE}", File.read(content_path) )
       File.open(tests_path, "w") { |file| file.puts tests }
     end
 
     def rename_template_files
-      `mv POD_README.md README.md`
-      `mv POD_LICENSE LICENSE`
-      `mv NAME.podspec #{pod_name}.podspec`
+      FileUtils.mv "POD_README.md", "README.md"
+      FileUtils.mv "POD_LICENSE", "LICENSE"
+      FileUtils.mv "NAME.podspec", "#{pod_name}.podspec"
     end
 
     def reinitialize_git_repo
       `rm -rf .git`
       `git init`
       `git add -A`
-      `git commit -m "Initial commit"`
     end
 
     def validate_user_details
